requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
_tflow_select=2.3.0=mkl
absl-py=0.8.1=py37_0
affinegap=1.10=pypi_0
appdirs=1.4.3=py37h28b3542_0
asn1crypto=1.2.0=py37_0
astor=0.8.0=py37_0
atomicwrites=1.3.0=pypi_0
attrs=19.3.0=py_0
autopep8=1.4.4=py_0
backcall=0.1.0=py37_0
binutils_impl_linux-64=2.31.1=h6176602_1
binutils_linux-64=2.31.1=h6176602_8
black=19.10b0=py37_0
blas=2.14=openblas
bleach=3.1.0=py37_0
boto=2.49.0=pypi_0
boto3=1.9.243=pypi_0
botocore=1.12.243=pypi_0
bpemb=0.3.0=pypi_0
btrees=4.6.0=pypi_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2019.10.16=0
categorical-distance=1.9=pypi_0
certifi=2019.9.11=py37_0
cffi=1.13.2=py37h8022711_0
chardet=3.0.4=py37_1003
click=7.0=py37_0
cryptography=2.8=py37h1ba5d50_0
cudatoolkit=10.0.130=0
cudnn=7.6.4=cuda10.0_0
cycler=0.10.0=py37_0
cymem=2.0.2=py37hfd86e86_0
cython-blis=0.4.1=py37h516909a_0
datetime-distance=0.1.3=pypi_0
dbus=1.13.6=he372182_0
decorator=4.4.1=py_0
dedupe=1.9.9=pypi_0
dedupe-hcluster=0.3.6=pypi_0
dedupe-variable-datetime=0.1.5=pypi_0
defusedxml=0.6.0=py_0
deprecated=1.2.6=pypi_0
docutils=0.15.2=pypi_0
doublemetaphone=0.1=pypi_0
en-core-web-lg=2.2.0=pypi_0
en-core-web-sm=2.2.0=pypi_0
entrypoints=0.3=py37_0
et_xmlfile=1.0.1=py37_0
expat=2.2.6=he6710b0_0
fastcluster=1.1.25=pypi_0
flair=0.4.3=pypi_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.0=he983fc9_1
future=0.17.1=pypi_0
gast=0.2.2=py37_0
gcc_impl_linux-64=7.3.0=habb00fd_1
gcc_linux-64=7.3.0=h553295d_8
gensim=3.8.1=pypi_0
gettext=0.19.8.1=hd7bead4_3
glib=2.58.3=h6f030ca_1002
google-pasta=0.1.7=py_0
grpcio=1.23.0=py37he9ae1f9_0
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
gxx_impl_linux-64=7.3.0=hdf63c60_1
gxx_linux-64=7.3.0=h553295d_8
h5py=2.10.0=nompi_py37h513d04c_100
haversine=2.1.2=pypi_0
hdf5=1.10.5=nompi_h3c11f04_1104
highered=0.2.1=pypi_0
hyperopt=0.1.2=pypi_0
icu=64.2=he1b5a44_1
idna=2.8=pypi_0
importlib_metadata=0.23=py37_0
intel-openmp=2019.5=281
ipykernel=5.1.3=py37h39e3cac_0
ipython=7.6.1=pypi_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
isort=4.3.21=py37_0
jdcal=1.4.1=py_0
jedi=0.15.1=py37_0
jinja2=2.10.3=py_0
jmespath=0.9.4=pypi_0
joblib=0.14.0=py_0
jpeg=9c=h14c3975_1001
json5=0.8.5=py_0
jsonschema=3.1.1=py37_0
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.0=py37_0
jupyterlab=1.2.1=py_0
jupyterlab_code_formatter=0.7.0=py_0
jupyterlab_server=1.0.6=py_0
keras=2.3.1=py37_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37he6710b0_0
langdetect=1.0.7=pypi_0
levenshtein-search=1.4.4=pypi_0
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libclang=9.0.0=hc9558a2_1
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libgpuarray=0.7.6=h14c3975_0
libiconv=1.15=h63c8f33_5
liblapack=3.8.0=14_openblas
liblapacke=3.8.0=14_openblas
libllvm9=9.0.0=hc9558a2_2
libopenblas=0.3.7=h6e990d7_3
libpng=1.6.37=hbc83047_0
libprotobuf=3.10.1=h8b12597_0
libsodium=1.0.17=h516909a_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=hfc65ed5_0
libuuid=2.32.1=h14c3975_1000
libuv=1.33.1=h516909a_0
libxcb=1.13=h1bed415_1
libxkbcommon=0.9.1=hebb1f50_0
libxml2=2.9.10=hee79883_0
lz4-c=1.8.3=he1b5a44_1001
mako=1.1.0=py_0
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.1=py37_2
matplotlib-base=3.1.1=py37h250f245_2
mistune=0.8.4=py37h7b6447c_0
mkl=2019.5=281
mkl-service=2.3.0=py37he904b0f_0
more-itertools=7.2.0=py37_0
mpld3=0.3=pypi_0
murmurhash=1.0.2=py37he6710b0_0
mypy_extensions=0.4.3=py37_0
nb-black=1.0.6=pypi_0
nbconvert=5.6.1=py37_0
nbformat=4.4.0=py37_0
ncurses=6.1=he6710b0_1
networkx=2.3=pypi_0
ninja=1.9.0=py37hfd86e86_0
nodejs=12.13.0=h10a4023_0
notebook=6.0.1=py37_0
nspr=4.23=he1b5a44_0
nss=3.47=he751ad9_0
numpy=1.17.3=py37h95a1406_0
olefile=0.46=py37_0
openpyxl=3.0.0=py_0
openssl=1.1.1=h7b6447c_0
opt_einsum=3.1.0=py_0
packaging=19.2=py_0
pandas=0.25.3=py37hb3f55d8_0
pandoc=2.7.3=0
pandocfilters=1.4.2=py37_1
parso=0.5.1=py_0
pathspec=0.6.0=py_0
patsy=0.5.1=py37_0
pcre=8.43=he6710b0_0
persistent=4.5.0=pypi_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.2.1=py37h6b7be26_0
pip=19.3.1=py37_0
plac=0.9.6=py37_0
pluggy=0.13.0=pypi_0
preshed=3.0.2=py37he1b5a44_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.10=py_0
protobuf=3.10.1=py37he1b5a44_0
ptyprocess=0.6.0=py37_0
py=1.8.0=pypi_0
pycodestyle=2.5.0=py37_0
pycparser=2.19=py37_0
pygments=2.4.2=py_0
pygpu=0.7.6=py37h035aef0_0
pyhacrf-datamade=0.2.3=pypi_0
pylbfgs=0.2.0.12=pypi_0
pymongo=3.9.0=pypi_0
pyopenssl=19.0.0=py37_0
pyparsing=2.4.4=py_0
pyqt=5.12.3=py37hcca6a23_0
pyqt5-sip=4.19.18=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.15.5=py37h516909a_0
pysocks=1.7.1=py37_0
pytest=5.2.1=pypi_0
python=3.7.5=h0371630_0
python-dateutil=2.8.1=py_0
pytorch=1.2.0=py3.7_cuda10.0.130_cudnn7.6.2_0
pytorch-transformers=1.2.0=pypi_0
pytz=2019.3=py_0
pyyaml=5.1.2=py37h7b6447c_0
pyzmq=18.1.0=py37he6710b0_0
qt=5.12.5=hd8c4c69_1
readline=7.0=h7b6447c_5
regex=2019.8.19=pypi_0
requests=2.22.0=py37_1
rlr=2.4.5=pypi_0
s3transfer=0.2.1=pypi_0
sacremoses=0.0.35=pypi_0
scikit-learn=0.21.3=py37h22eb022_0
scipy=1.3.1=py37he2b7bc3_0
seaborn=0.9.0=py37_0
segtok=1.5.7=pypi_0
send2trash=1.5.0=py37_0
sentencepiece=0.1.83=pypi_0
setuptools=41.6.0=py37_0
simplecosine=1.2=pypi_0
simplejson=3.16.0=pypi_0
sip=4.19.18=py37he1b5a44_1
six=1.13.0=py37_0
sklearn=0.0=pypi_0
smart-open=1.8.4=pypi_0
spacy=2.2.2=py37hc9558a2_0
sqlite=3.30.1=h7b6447c_0
sqlitedict=1.6.0=pypi_0
srsly=0.2.0=py37he1b5a44_0
statsmodels=0.10.1=py37hdd07704_0
tabulate=0.8.5=pypi_0
tensorboard=2.0.0=pyhb230dea_0
tensorflow=2.0.0=mkl_py37h66b46cc_0
tensorflow-base=2.0.0=mkl_py37h9204916_0
tensorflow-estimator=2.0.0=pyh2649769_0
termcolor=1.1.0=py37_1
terminado=0.8.2=py37_0
testpath=0.4.4=py_0
theano=1.0.4=py37hfd86e86_0
thinc=7.3.0=py37hc9558a2_0
tk=8.6.9=hed695b0_1003
toml=0.10.0=py37h28b3542_0
torchvision=0.4.0=cuda100py37hecfc37a_0
tornado=6.0.3=py37h7b6447c_0
tqdm=4.36.1=py_0
traitlets=4.3.3=py37_0
typed-ast=1.4.0=py37h7b6447c_0
typing_extensions=3.7.4.1=py37_0
unidecode=1.1.1=py_0
urllib3=1.24.3=pypi_0
wasabi=0.4.0=py_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.11.2=py37h7b6447c_0
xlrd=1.2.0=py37_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
yapf=0.28.0=py_0
zeromq=4.3.2=he1b5a44_2
zipp=0.6.0=py_0
zlib=1.2.11=h7b6447c_3
zope-index=4.4.0=pypi_0
zope-interface=4.6.0=pypi_0
zstd=1.4.3=h3b9ef0a_0
